+++ 3rdparty/zlib/zlib.cmake
@@ -10,17 +10,17 @@ find_package(Git QUIET REQUIRED)
 
 ExternalProject_Add(
     ext_zlib
-    PREFIX zlib
-    URL https://github.com/madler/zlib/archive/refs/tags/v1.2.13.tar.gz
-    URL_HASH SHA256=1525952a0a567581792613a9723333d7f8cc20b87a81f920fb8bc7e3f2251428
+    PREFIX zlib-ng
+    URL https://github.com/zlib-ng/zlib-ng/archive/refs/tags/2.2.2.tar.gz
+    URL_HASH SHA256=fcb41dd59a3f17002aeb1bb21f04696c9b721404890bb945c5ab39d2cb69654c
     DOWNLOAD_DIR "${OPEN3D_THIRD_PARTY_DOWNLOAD_DIR}/zlib"
-    UPDATE_COMMAND ""
-    PATCH_COMMAND ${GIT_EXECUTABLE} init
-    COMMAND       ${GIT_EXECUTABLE} apply --ignore-space-change --ignore-whitespace
-                  ${CMAKE_CURRENT_LIST_DIR}/0001-patch-zlib-to-enable-unzip.patch
     CMAKE_ARGS
         -DCMAKE_POLICY_VERSION_MINIMUM=3.5
         -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>
+        -DZLIB_COMPAT:BOOL=ON
+        -DZLIB_ENABLE_TESTS:BOOL=OFF
+        -DZLIBNG_ENABLE_TESTS:BOOL=OFF
+        -DWITH_GTEST:BOOL=OFF
         # zlib needs visiible symbols for examples. Disabling example building causes
         # assember error in GPU CI. zlib symbols are hidden during linking.
         ${ExternalProject_CMAKE_ARGS}